refactor:清理alice无效测试
