+++ tests/unit/alice/orchestration/test_frame_factory.py
@@ -3,8 +3,6 @@ from __future__ import annotations
 from hivememory.agent_runtime.policy import FrameExecutionPolicy
 from hivememory.alice.orchestration.frame_factory import FrameFactory, FrameSpec
 from hivememory.core.models import OMNI_DOLL_PROFILE, Identity
-from hivememory.prompts.assembler import AgentPromptAssembler
-from hivememory.system.config import KoakumaConfig
 
 
 def test_frame_factory_creates_root_frame_from_spec() -> None:
@@ -20,10 +18,6 @@ def test_frame_factory_creates_root_frame_from_spec() -> None:
         )
     )
 
-    assert frame.runtime_scope.run_id == "run-1"
-    assert frame.runtime_scope.frame_id == "frame-1"
-    assert frame.topic_id == "topic-1"
-    assert frame.working_history == [{"role": "system", "content": "hello"}]
     assert frame.progress.turn_events == []
     assert frame.progress.sequence == 0
 
@@ -45,7 +39,6 @@ def test_frame_factory_creates_transient_frame_with_same_run_id() -> None:
     )
 
     assert frame.is_transient()
-    assert frame.runtime_scope.run_id == "run-1"
     assert frame.execution_policy.allows("READ")
     assert not frame.execution_policy.allows("CALL")
     assert [event.kind for event in frame.progress.turn_events] == ["user_message"]
@@ -74,7 +67,6 @@ def test_frame_factory_records_only_latest_user_message() -> None:
         )
     )
 
-    assert frame.working_history == messages
     assert len(frame.progress.turn_events) == 1
     assert frame.progress.turn_events[0].content == "current"
     assert frame.progress.turn_events[0].sequence == 0
@@ -99,15 +91,3 @@ def test_frame_factory_does_not_record_empty_latest_user_message() -> None:
 
     assert frame.progress.turn_events == []
     assert frame.progress.sequence == 0
-
-
-def test_sub_agent_prompt_disables_call() -> None:
-    assembler = AgentPromptAssembler(KoakumaConfig())
-
-    messages = assembler.build_sub_agent_messages(
-        profile=OMNI_DOLL_PROFILE,
-        task="Write unit tests",
-    )
-
-    assert "CALL" not in messages[0]["content"]
-    assert "READ" in messages[0]["content"]
