+++ tests/unit/alice/orchestration/test_run_executor.py
@@ -16,7 +16,7 @@ from hivememory.alice.orchestration.sub_agent.call_coordinator import (
     DispatchCallee,
     ResumeCaller,
 )
-from hivememory.alice.runtime.streaming import AgentRunStream, QueueAgentRunOutput
+from hivememory.alice.runtime.streaming import QueueAgentRunOutput
 from hivememory.core.mtp import MTPCallRequest
 
 
@@ -120,45 +120,6 @@ async def test_queue_sink_applies_backpressure_at_capacity():
 
     await second_emit
     assert (await output.receive())["data"]["stream_sequence"] == 1
-
-
-@pytest.mark.asyncio
-async def test_run_executor_reenters_suspended_frame_with_continuous_stream_sequence():
-    calls = 0
-
-    async def run_frame(_frame, *, output_sink, **_kwargs):
-        nonlocal calls
-        calls += 1
-        await output_sink.send(TokenDelta(content=str(calls)))
-        if calls == 1:
-            return FrameExecutionResult(
-                status=FrameExecutionStatus.SUSPENDED,
-                call_request=MTPCallRequest(target_alias="helper", task="task"),
-                suspend_action_id="act-1",
-            )
-        return FrameExecutionResult(status=FrameExecutionStatus.COMPLETED)
-
-    frame = _frame_stub("frame-1")
-    session = _session_with(frame)
-    coordinator = _CallCoordinatorStub(emit_end=True)
-    executor = RunExecutor(
-        SimpleNamespace(run_frame=run_frame, apply_call_response=MagicMock()),
-        session=session,
-        call_coordinator=coordinator,
-    )
-    agent_stream = AgentRunStream(session)
-    events = [
-        event
-        async for event in agent_stream.events(executor.run(frame, run_output=agent_stream.output))
-    ]
-
-    assert [event["data"]["stream_sequence"] for event in events] == [0, 1, 2, 3]
-    assert [event["data"]["agent_run_id"] for event in events] == ["run-1"] * 4
-    assert agent_stream.next_sequence == 4
-    assert executor.terminal_result is not None
-    assert executor.terminal_result.status == FrameExecutionStatus.COMPLETED
-    assert set(session.frames) == {"frame-1", "frame-child"}
-    assert session.call_records[("frame-1", "act-1")].status.value == "applied"
 
 
 @pytest.mark.asyncio
@@ -361,111 +322,6 @@ async def test_run_executor_preserves_cancellation_when_call_cleanup_fails():
 
 
 @pytest.mark.asyncio
-async def test_run_executor_cancels_runner_when_stream_consumer_closes():
-    runner_cancelled = asyncio.Event()
-
-    async def run_frame(_frame, *, output_sink, **_kwargs):
-        try:
-            await output_sink.send(TokenDelta(content="started"))
-            await asyncio.Event().wait()
-        finally:
-            runner_cancelled.set()
-
-    frame = _frame_stub("frame-1")
-    session = _session_with(frame)
-    executor = RunExecutor(SimpleNamespace(run_frame=run_frame), session=session)
-    agent_stream = AgentRunStream(session)
-    stream = agent_stream.events(executor.run(frame, run_output=agent_stream.output))
-
-    first_event = await anext(stream)
-    await stream.aclose()
-    await asyncio.wait_for(runner_cancelled.wait(), timeout=1)
-
-    assert first_event["event"] == "token"
-    assert executor.terminal_result is not None
-    assert executor.terminal_result.status == FrameExecutionStatus.CANCELLED
-
-
-@pytest.mark.asyncio
-async def test_run_executor_cleans_active_call_when_stream_consumer_closes():
-    root_calls = 0
-
-    async def run_frame(frame, *, output_sink, **_kwargs):
-        nonlocal root_calls
-        if frame.runtime_scope.frame_id == "frame-child":
-            await output_sink.send(TokenDelta(content="child"))
-            await asyncio.Event().wait()
-        root_calls += 1
-        await output_sink.send(TokenDelta(content="root"))
-        return FrameExecutionResult(
-            status=FrameExecutionStatus.SUSPENDED,
-            call_request=MTPCallRequest(target_alias="helper", task="task"),
-            suspend_action_id="act-1",
-        )
-
-    finalize_run = MagicMock()
-    frame = _frame_stub("frame-1")
-    session = _session_with(frame)
-    coordinator = _CallCoordinatorStub()
-    executor = RunExecutor(
-        SimpleNamespace(run_frame=run_frame, finalize_run=finalize_run),
-        session=session,
-        call_coordinator=coordinator,
-    )
-    agent_stream = AgentRunStream(session)
-    stream = agent_stream.events(executor.run(frame, run_output=agent_stream.output))
-
-    assert (await anext(stream))["data"]["content"] == "root"
-    assert (await anext(stream))["data"]["content"] == "child"
-    await stream.aclose()
-
-    assert coordinator.cancel_calls == 1
-    assert session.call_records[("frame-1", "act-1")].status.value == "cancelled"
-    finalize_run.assert_called_once()
-    assert finalize_run.call_args.args[1].status == FrameExecutionStatus.CANCELLED
-
-
-@pytest.mark.asyncio
-async def test_run_executor_cancels_record_during_call_preparation_await():
-    preparation_started = asyncio.Event()
-
-    class BlockingPreparationCoordinator(_CallCoordinatorStub):
-        async def begin_call(self, caller, suspension, *, session, **_kwargs):
-            record = session.register_call(caller, suspension.suspend_action_id)
-            record.begin_resolution()
-            preparation_started.set()
-            await asyncio.Event().wait()
-
-    async def run_frame(_frame, *, output_sink, **_kwargs):
-        await output_sink.send(TokenDelta(content="root"))
-        return FrameExecutionResult(
-            status=FrameExecutionStatus.SUSPENDED,
-            call_request=MTPCallRequest(target_alias="helper", task="task"),
-            suspend_action_id="act-1",
-        )
-
-    finalize_run = MagicMock()
-    frame = _frame_stub("frame-1")
-    session = _session_with(frame)
-    coordinator = BlockingPreparationCoordinator()
-    executor = RunExecutor(
-        SimpleNamespace(run_frame=run_frame, finalize_run=finalize_run),
-        session=session,
-        call_coordinator=coordinator,
-    )
-    agent_stream = AgentRunStream(session)
-    stream = agent_stream.events(executor.run(frame, run_output=agent_stream.output))
-
-    await anext(stream)
-    await asyncio.wait_for(preparation_started.wait(), timeout=1)
-    await stream.aclose()
-
-    assert coordinator.cancel_calls == 1
-    assert session.call_records[("frame-1", "act-1")].status.value == "cancelled"
-    finalize_run.assert_called_once()
-
-
-@pytest.mark.asyncio
 async def test_run_executor_accepts_call_resume_without_applying_response_itself():
     calls = 0
 
@@ -516,7 +372,6 @@ async def test_run_executor_finalizes_each_terminal_status_exactly_once(status):
 
     result = await executor.run(frame)
 
-    assert result.status == status
     finalize_run.assert_called_once_with("run-1", result)
