+++ tests/unit/alice/test_system.py
@@ -1,37 +1,14 @@
+"""
+AliceSystem 单元测试（unit 保留集）
+
+系统 facade 对全局总线路由挂载/卸载的真实协作测试已迁移至
+tests/integration/alice/test_system.py。本文件保留 health 报告测试。
+"""
+
 import pytest
 
-from hivememory.alice.contracts.public_routes import AliceRoutes
 from hivememory.alice.system import AliceSystem
 from hivememory.system.config import HiveMemoryConfig
-from hivememory.system.runtime.bus.global_bus import GlobalSystemBus
-
-
-@pytest.mark.asyncio
-async def test_start_registers_public_routes_and_stop_unregisters():
-    bus = GlobalSystemBus()
-    system = AliceSystem(config=HiveMemoryConfig(), global_bus=bus)
-
-    await system.start()
-
-    assert AliceRoutes.RUN_AGENT in bus.list_routes()
-    assert AliceRoutes.RUN_AGENT_STREAM in bus.list_routes()
-
-    await system.stop()
-
-    assert AliceRoutes.RUN_AGENT not in bus.list_routes()
-    assert AliceRoutes.RUN_AGENT_STREAM not in bus.list_routes()
-
-
-@pytest.mark.asyncio
-async def test_start_is_idempotent_for_public_routes():
-    bus = GlobalSystemBus()
-    system = AliceSystem(config=HiveMemoryConfig(), global_bus=bus)
-
-    await system.start()
-    await system.start()
-
-    assert bus.list_routes().count(AliceRoutes.RUN_AGENT) == 1
-    await system.stop()
 
 
 @pytest.mark.asyncio
@@ -40,5 +17,5 @@ async def test_health_reports_runtime_health():
 
     health = await system.health()
 
-    assert health["status"] == "ok"
-    assert "runtime" in health
+    assert health["runtime"]["koakuma_runtime"]["status"] == "ok"
+    assert "agent_runtime" in health["runtime"]
